requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
llama-index-core>=0.10.0
llama-index-llms-gemini>=0.1.0
pydantic==2.3.0
python-dotenv==1.0.0
pyyaml>=6.0.1
